Applied Eclipse's cleanup function on all projects with very conservative settings.
git-svn-id: https://svn.jfire.org/svn/main/trunk/JFireTrade@9054 18c61ebf-f106-0410-9e1e-81fb82415b8f

+++ src/org/nightlabs/jfire/accounting/Currency.java
@@ -139,6 +139,6 @@ implements Serializable, org.nightlabs.l10n.Currency
 	 * @return
 	 */
 	public double toDouble(long amount) {		
-		return (double)amount / Math.pow(10, getDecimalDigitCount());		
+		return amount / Math.pow(10, getDecimalDigitCount());		
 	}
 }
